Refactored example to Helidon MP
