+++ examples/security/attribute-based-access-control/pom.xml
@@ -39,30 +39,9 @@
 
     <dependencies>
         <dependency>
-            <groupId>io.helidon.security.integration</groupId>
-            <artifactId>helidon-security-integration-jersey</artifactId>
+            <groupId>io.helidon.microprofile.bundles</groupId>
+            <artifactId>helidon-microprofile-1.2</artifactId>
             <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.helidon.bundles</groupId>
-            <artifactId>helidon-bundles-jersey</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-         <dependency>
-             <groupId>io.helidon.bundles</groupId>
-             <artifactId>helidon-bundles-config</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.helidon.bundles</groupId>
-            <artifactId>helidon-bundles-security</artifactId>
-            <version>${project.version}</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>com.google.code.findbugs</groupId>
-                    <artifactId>jsr305</artifactId>
-                </exclusion>
-            </exclusions>
         </dependency>
         <dependency>
             <groupId>org.glassfish</groupId>
