+++ examples/security/attribute-based-access-control/src/main/java/io/helidon/security/examples/abac/AbacApplication.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2019 Oracle and/or its affiliates. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.helidon.security.examples.abac;
+
+import java.util.Set;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.ws.rs.ApplicationPath;
+import javax.ws.rs.core.Application;
+
+import io.helidon.common.CollectionsHelper;
+
+/**
+ * Application class of this MP application.
+ */
+@ApplicationScoped
+@ApplicationPath("/rest")
+public class AbacApplication extends Application {
+    @Override
+    public Set<Class<?>> getClasses() {
+        return CollectionsHelper.setOf(AbacResource.class, AbacExplicitResource.class);
+    }
+}
