+++ examples/security/attribute-based-access-control/src/main/resources/application.yaml
@@ -18,46 +18,47 @@
 # It is possible to provide default config values, here.
 #
 
-webserver:
+server:
   port: 8380
 
 security:
   providers:
     - abac:
-      # prepares environment
-      # executes attribute validations
-      # validates that attributes were processed
-      # grants/denies access to resource
-      #
-      ####
-      # Combinations:
-      # # Will fail if any attribute is not validated and if any has failed validation
-      # fail-on-unvalidated: true
-      # fail-if-none-validated: true
-      #
-      # # Will fail if there is one or more attributes present and NONE of them is validated or if any has failed validation
-      # # Will NOT fail if there is at least one validated attribute and any number of not validated attributes (and NONE failed)
-      # fail-on-unvalidated: false
-      # fail-if-none-validated: true
-      #
-      # # Will fail if there is any attribute that failed validation
-      # # Will NOT fail if there are no failed validation or if there are NONE validated
-      # fail-on-unvalidated: false
-      # fail-if-none-validated: false
-      ####
-      # fail if an attribute was not validated (e.g. we do not know, whether it is valid or not)
-      # defaults to true
-      fail-on-unvalidated: true
-      # fail if none of the attributes were validated
-      # defaults to true
-      fail-if-none-validated: true
-      policy-validator:
-        validators:
-        - class: "io.helidon.security.abac.policy.DefaultPolicyValidator"
-          - class: "..."
-        my-custom-policy-engine:
-          some-key: "some value"
-          another-key: "another value"
+        # prepares environment
+        # executes attribute validations
+        # validates that attributes were processed
+        # grants/denies access to resource
+        #
+        ####
+        # Combinations:
+        # # Will fail if any attribute is not validated and if any has failed validation
+        # fail-on-unvalidated: true
+        # fail-if-none-validated: true
+        #
+        # # Will fail if there is one or more attributes present and NONE of them is validated or if any has failed validation
+        # # Will NOT fail if there is at least one validated attribute and any number of not validated attributes (and NONE failed)
+        # fail-on-unvalidated: false
+        # fail-if-none-validated: true
+        #
+        # # Will fail if there is any attribute that failed validation
+        # # Will NOT fail if there are no failed validation or if there are NONE validated
+        # fail-on-unvalidated: false
+        # fail-if-none-validated: false
+        ####
+        # fail if an attribute was not validated (e.g. we do not know, whether it is valid or not)
+        # defaults to true
+        fail-on-unvalidated: true
+        # fail if none of the attributes were validated
+        # defaults to true
+        fail-if-none-validated: true
+#        policy-validator:
+#          validators:
+#          - class: "io.helidon.security.abac.policy.DefaultPolicyValidator"
+#          my-custom-policy-engine:
+#            some-key: "some value"
+#            another-key: "another value"
+    - atn:
+      class: "io.helidon.security.examples.abac.AtnProvider"
   web-server:
     defaults:
     paths:
